// src/graphql/type.interface.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};

export enum ActiveCodeEnum {
  ACTIVATE = 'ACTIVATE',
  FORGOT = 'FORGOT',
}

export enum ActorTypeEnum {
  ADMIN = 'ADMIN',
  PARTNER = 'PARTNER',
  USER = 'USER',
}

export type AddressEntity = Node & {
  __typename?: 'AddressEntity';
  addressDetail?: Maybe<Scalars['String']>;
  addressName: Scalars['String'];
  contactName: Scalars['String'];
  contactPhone: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  isDefault: Scalars['Boolean'];
  latitude: Scalars['Float'];
  longitude: Scalars['Float'];
  mapAddress: Scalars['String'];
  updatedAt: Scalars['DateTime'];
  userId: Scalars['String'];
};
// ...
export type AdminEntity = Node & {
  __typename?: 'AdminEntity';
  address?: Maybe<Scalars['String']>;
  addressMoreInfo?: Maybe<Scalars['String']>;
  avatar?: Maybe<Media>;
  avatarId?: Maybe<Scalars['String']>;
  birthday?: Maybe<Scalars['DateTime']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  email: Scalars['String'];
  fullname?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  type: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type CartItemEntity = Node & {
  __typename?: 'CartItemEntity';
  cartId: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  product: ProductEntity;
  productId: Scalars['String'];
  quantity: Scalars['Float'];
  store?: Maybe<PartnerEntity>;
  total: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
};

export type CartItemInput = {
  productId: Scalars['String'];
  quantity: Scalars['Float'];
};

export type CategoryConnection = {
  __typename?: 'CategoryConnection';
  items?: Maybe<Array<CategoryEntity>>;
  meta: BasePaginationMeta;
};

export type CategoryEntity = Node & {
  __typename?: 'CategoryEntity';
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  name: Scalars['String'];
  type: CategoryTypeEnum;
  updatedAt: Scalars['DateTime'];
};

export enum CategoryTypeEnum {
  CANCEL_MAINTENANCE_REASON = 'CANCEL_MAINTENANCE_REASON',
  CANCEL_ORDER_REASON_BY_PARTNER = 'CANCEL_ORDER_REASON_BY_PARTNER',
  CANCEL_ORDER_REASON_BY_USER = 'CANCEL_ORDER_REASON_BY_USER',
  CANCEL_QUOTATION_REASON = 'CANCEL_QUOTATION_REASON',
  CANCEL_REASON = 'CANCEL_REASON',
  CANCEL_REASON_BY_PARTNER = 'CANCEL_REASON_BY_PARTNER',
  EDUCATION = 'EDUCATION',
  LEVEL = 'LEVEL',
  MANUFACTURER = 'MANUFACTURER',
  MODEL = 'MODEL',
  ORIGIN = 'ORIGIN',
  PART_OF_PRODUCT = 'PART_OF_PRODUCT',
  PROBLEM = 'PROBLEM',
  PRODUCT_UNIT = 'PRODUCT_UNIT',
  QUALIFICATION = 'QUALIFICATION',
  QUOTATION_PRICE_LIST = 'QUOTATION_PRICE_LIST',
  STORE_PRODUCT = 'STORE_PRODUCT',
  VEHICLE_TYPE = 'VEHICLE_TYPE',
}
// ...
export type DiscountCodeEntity = Node & {
  __typename?: 'DiscountCodeEntity';
  admin?: Maybe<AdminEntity>;
  adminId?: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  endDate?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  isAssignAllProduct: Scalars['Boolean'];
  limit?: Maybe<Scalars['Float']>;
  limitPerAccount?: Maybe<Scalars['Float']>;
  minOrderValue: Scalars['Float'];
  name: Scalars['String'];
  partner?: Maybe<PartnerEntity>;
  partnerId?: Maybe<Scalars['String']>;
  products?: Maybe<Array<ProductEntity>>;
  startDate: Scalars['DateTime'];
  unit: DiscountCodeUnitEnum;
  updatedAt: Scalars['DateTime'];
  usedCount: Scalars['Float'];
  value: Scalars['Float'];
};

export enum DiscountCodeUnitEnum {
  PERCENTAGE = 'PERCENTAGE',
  VND = 'VND',
}
// ...
export type Expense = {
  __typename?: 'Expense';
  cost?: Maybe<Scalars['Float']>;
  distance?: Maybe<Scalars['Float']>;
  time?: Maybe<Scalars['Float']>;
};
// ...
export enum FileType {
  EXCEL = 'EXCEL',
  FILE = 'FILE',
  IMAGE = 'IMAGE',
  INVOICE = 'INVOICE',
  VIDEO = 'VIDEO',
}
// ...
export type Media = Node & {
  __typename?: 'Media';
  createdAt: Scalars['DateTime'];
  fileSize?: Maybe<Scalars['Int']>;
  fullOriginalUrl?: Maybe<Scalars['String']>;
  fullThumbUrl?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isDeleted: Scalars['Boolean'];
  mimeType?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  originalUrl: Scalars['String'];
  ownerId?: Maybe<Scalars['String']>;
  thumbUrl?: Maybe<Scalars['String']>;
  type: FileType;
  updatedAt: Scalars['DateTime'];
  videoUrl?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export enum OperatingUnitEnum {
  HOURS = 'HOURS',
  KM = 'KM',
}
// ...
export type PartnerEntity = Node & {
  __typename?: 'PartnerEntity';
  addressMoreInfo?: Maybe<Scalars['String']>;
  avatar?: Maybe<Media>;
  avatarId?: Maybe<Scalars['String']>;
  bank: Scalars['String'];
  birthday?: Maybe<Scalars['DateTime']>;
  cardNumber: Scalars['String'];
  citizenId?: Maybe<Scalars['String']>;
  countProduct?: Maybe<Scalars['Int']>;
  countTechnician?: Maybe<Scalars['Int']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  education?: Maybe<CategoryEntity>;
  email?: Maybe<Scalars['String']>;
  expenseInfo?: Maybe<Expense>;
  fullname?: Maybe<Scalars['String']>;
  hotline?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  isAdmin: Scalars['Boolean'];
  isApproved?: Maybe<Scalars['Boolean']>;
  latitude?: Maybe<Scalars['Float']>;
  level?: Maybe<CategoryEntity>;
  longitude?: Maybe<Scalars['Float']>;
  mapAddress?: Maybe<Scalars['String']>;
  menus?: Maybe<Array<PartnerMenuEnum>>;
  parentId?: Maybe<Scalars['String']>;
  parentInfo?: Maybe<PartnerEntity>;
  phone: Scalars['String'];
  qualifications?: Maybe<Array<CategoryEntity>>;
  reviewSummary?: Maybe<ReviewSummary>;
  star: Scalars['Float'];
  starInfo: Array<StarInfo>;
  storeReviewSummary?: Maybe<ReviewSummary>;
  storeStar?: Maybe<Scalars['Float']>;
  storeStarInfo?: Maybe<Array<StarInfo>>;
  suggestionPoint: Scalars['Float'];
  type: PartnerTypeEnum;
  updatedAt: Scalars['DateTime'];
};

export type PartnerForgotPasswordInput = {
  emailOrPhone: Scalars['String'];
};

export type PartnerLoginInput = {
  deviceId?: InputMaybe<Scalars['String']>;
  email?: InputMaybe<Scalars['String']>;
  password: Scalars['String'];
  phone?: InputMaybe<Scalars['String']>;
};

export enum PartnerMenuEnum {
  COURSE = 'COURSE',
  DISCOUNT = 'DISCOUNT',
  DOCUMENT = 'DOCUMENT',
  ORDER = 'ORDER',
  PRODUCT = 'PRODUCT',
  REPAIR = 'REPAIR',
  TECHNICIAN = 'TECHNICIAN',
}
// ...
export enum PartnerTypeEnum {
  AGENCY = 'AGENCY',
  FREELANCER_TECHNICIAN = 'FREELANCER_TECHNICIAN',
  TECHNICIAN = 'TECHNICIAN',
}
// ...
export type ProductDevice = {
  __typename?: 'ProductDevice';
  manufacturer: CategoryEntity;
  model: CategoryEntity;
  vehicleType: CategoryEntity;
};

export type ProductDeviceInput = {
  manufacturerId: Scalars['String'];
  modelId: Scalars['String'];
  vehicleTypeId: Scalars['String'];
};

export type ProductEntity = Node & {
  __typename?: 'ProductEntity';
  avatar?: Maybe<Media>;
  avatarId: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  descriptionImageIds?: Maybe<Array<Scalars['String']>>;
  descriptionImages: Array<Media>;
  detail?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  isFavorite?: Maybe<Scalars['Boolean']>;
  isFixedCost: Scalars['Boolean'];
  isNew: Scalars['Boolean'];
  manufacturer?: Maybe<CategoryEntity>;
  model?: Maybe<CategoryEntity>;
  name: Scalars['String'];
  numberOrder?: Maybe<Scalars['Float']>;
  numberSold: Scalars['Float'];
  operatingNumber?: Maybe<Scalars['Float']>;
  operatingUnit?: Maybe<OperatingUnitEnum>;
  ordinalNumber?: Maybe<Scalars['Float']>;
  origin?: Maybe<CategoryEntity>;
  partNumber?: Maybe<Scalars['String']>;
  partOfProduct?: Maybe<CategoryEntity>;
  partner?: Maybe<PartnerEntity>;
  partnerId?: Maybe<Scalars['String']>;
  productDevices: Array<ProductDevice>;
  productType?: Maybe<CategoryEntity>;
  productUnit?: Maybe<CategoryEntity>;
  quantity: Scalars['Float'];
  reviewSummary?: Maybe<ReviewSummary>;
  serialNumber?: Maybe<Scalars['String']>;
  star: Scalars['Float'];
  starInfo?: Maybe<Array<StarInfo>>;
  type: ProductTypeEnum;
  unitPrice: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
  vehicleRegistrationPlate?: Maybe<Scalars['String']>;
  vinNumber?: Maybe<Scalars['String']>;
  yearOfManufacture?: Maybe<Scalars['Float']>;
};
// ...
export enum ProductTypeEnum {
  ACCESSARY = 'ACCESSARY',
  VEHICLE = 'VEHICLE',
}
// ...
export type ReviewSummary = {
  __typename?: 'ReviewSummary';
  percent: Scalars['Float'];
  starAverage: Scalars['Float'];
  total: Scalars['Float'];
};
// ...
export type StarInfo = {
  __typename?: 'StarInfo';
  star: Scalars['Float'];
  total: Scalars['Float'];
};